+++ Station/src/main/java/com/five/service/StationService.java
@@ -4,5 +4,5 @@ import java.math.BigDecimal;
 
 public interface StationService {
 
-    BigDecimal checkRoute(String sourceStation, String destinationStation);
+    double checkRoute(String sourceStation, String destinationStation);
 }
